+++ src/Game/libs/Snake/Systems/FruitSpawner.hpp
@@ -18,7 +18,7 @@ namespace Snake
             FruitSpawner() = default;
             ~FruitSpawner() override = default;
 
-            void run(float deltaTime,
+            void run(double deltaTime,
                 Arcade::ECS::IEventManager &eventManager,
                 Arcade::ECS::IEntityManager &entityManager) override;
